Group ManyPets into a feature folder
