+++ apps/store/src/features/manyPets/manyPets.constants.ts
@@ -1,0 +1,1 @@
+export const STORYBLOK_MANYPETS_FOLDER_SLUG = 'manypets'
