+++ apps/store/src/features/manyPets/manyPetsBlocks.ts
@@ -1,0 +1,3 @@
+import { ManyPetsMigrationPageBlock } from './ManyPetsMigrationPageBlock'
+
+export const manyPetsBlocks = [ManyPetsMigrationPageBlock]
